LayoutTests/imported/w3c/web-platform-tests/background-fetch/match.https.window-expected.txt


Harness Error (TIMEOUT), message = null

FAIL Matching to a single request should work assert_equals: expected "Background Fetch" but got ""
TIMEOUT Matching to a non-existing request should work Test timed out
NOTRUN Matching multiple times on the same request works as expected.
NOTRUN Access to active fetches is supported.
NOTRUN Match with query options.
